pyproject.toml
[project]
name = "document-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.72.0",
    "numpy>=2.4.2",
    "openai>=2.16.0",
    "pillow>=11.3.0",
    "pypdf>=6.6.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
